+++ src/component/Modal.js
@@ -1,6 +1,6 @@
 import { DOM } from "../utils/dom.js";
 import { $ } from "../utils/querySelectors.js";
-import AddLunchModalForm from "./AddLunchModalForm.js";
+import AddLunchModalForm from "./AddLunchModal/AddLunchModalForm.js";
 
 const Modal = {
   create(id, modalContent) {
